+++ framework/domain/src/main/java/is/codion/framework/domain/entity/EntitySerializer.java
@@ -112,6 +112,7 @@ final class EntitySerializer {
 	}
 
 	private static void serializeValues(DefaultEntity entity, ObjectOutputStream stream) throws IOException {
+		entity.clearDerivedCache();
 		serializeValues(entity.values, stream);
 		boolean modified = entity.originalValues != null;
 		stream.writeBoolean(modified);