+++ changelog.md
@@ -19,6 +19,8 @@ Codion Change Log
 - Controls.empty() now ignores separators.
 ### is.codion.framework.domain
 - Entity.referencedEntity() and referencedKey() renamed entity() and key().
+- Attribute.denormalized() added.
+- DefaultEntity now caches derived values.
 ### is.codion.swing.framework.model
 - SwingEntityTableModel.replace() now triggers multiple targeted table row updated events, instead of one for all rows.
 - SwingEntityTableModel.editEvents() removed, UpdateListener duplicated detail model link functionality by replacing foreign key values.